+++ src/js/common/schemaform/helpers.js
@@ -7,8 +7,8 @@ import shouldUpdate from 'recompose/shouldUpdate';
 import { deepEquals } from 'react-jsonschema-form/lib/utils';
 
 import { getInactivePages, getActivePages } from '../utils/helpers';
-import FormSaved from './FormSaved';
-import SaveInProgressErrorPage from './SaveInProgressErrorPage';
+import FormSaved from './save-in-progress/FormSaved';
+import SaveInProgressErrorPage from './save-in-progress/SaveInProgressErrorPage';
 
 export function createFormPageList(formConfig) {
   return Object.keys(formConfig.chapters)
