+++ src/js/common/schemaform/save-in-progress/savedFormRequest.js
@@ -1,6 +1,6 @@
 import merge from 'lodash/fp/merge';
 
-import environment from '../helpers/environment';
+import environment from '../../helpers/environment';
 
 function isJson(response) {
   const contentType = response.headers.get('content-type');
