+++ src/js/common/schemaform/state/index.js
@@ -1,6 +1,6 @@
 import {
   createInitialState,
-} from '../formState';
+} from '../state/helpers';
 
 import {
   createSaveInProgressInitialState,
