+++ src/js/pre-need/reducer.js
@@ -1,5 +1,5 @@
 import formConfig from './config/form';
-import createSchemaFormReducer from '../common/schemaform/reducers';
+import createSchemaFormReducer from '../common/schemaform/state';
 
 export default {
   form: createSchemaFormReducer(formConfig)
