+++ test/common/schemaform/SaveInProgressIntro.unit.spec.jsx
@@ -3,7 +3,7 @@ import moment from 'moment';
 import { expect } from 'chai';
 import SkinDeep from 'skin-deep';
 
-import SaveInProgressIntro from '../../../src/js/common/schemaform/SaveInProgressIntro';
+import SaveInProgressIntro from '../../../src/js/common/schemaform/save-in-progress/SaveInProgressIntro';
 
 describe('Schemaform <SaveInProgressIntro>', () => {
   const pageList = [{
