+++ test/util/schemaform-utils.jsx
@@ -11,7 +11,7 @@ import { fillDate as oldFillDate } from './unit-helpers';
 import {
   replaceRefSchemas,
   updateSchemaAndData
-} from '../../src/js/common/schemaform/formState';
+} from '../../src/js/common/schemaform/state/helpers';
 
 function getDefaultData(schema) {
   if (schema.type === 'array') {
